refactor: update chat client imports and remove deprecated types
- Refactored imports in chat client modules to utilize the new `interface.types` structure.
- Replaced instances of `ServerState` with `StateUpdate` in relevant classes to align with recent type changes.
- Removed the obsolete `types.py` file, consolidating type definitions for improved organization and clarity.
