+++ .gitignore
@@ -9,6 +9,9 @@ venv/
 __pycache__/
 **.egg-info/
 
+# Local cursor rules
+.cursor/rules/local/
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
