+++ packages/ragbits-chat/src/ragbits/chat/clients/sync_client.py
@@ -6,15 +6,9 @@ from typing import Any
 import httpx
 
 from .._utils import build_api_url, parse_sse_line
+from ..interface.types import ChatResponse, ChatResponseType, Message, MessageRole, StateUpdate
 from .base import SyncChatClientBase
 from .exceptions import ChatClientRequestError, ChatClientResponseError
-from .types import (
-    ChatResponse,
-    ChatResponseType,
-    Message,
-    MessageRole,
-    ServerState,
-)
 
 __all__ = ["RagbitsChatClient"]
 
@@ -38,7 +32,7 @@ class RagbitsChatClient(SyncChatClientBase):
 
         self.history: list[Message] = []
         self.conversation_id: str | None = None
-        self.server_state: ServerState | None = None
+        self.server_state: StateUpdate | None = None
 
         self._streaming_response: httpx.Response | None = None
